requirements.txt
loguru
huggingface-hub
python-dotenv